pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "efcd"
version = "0.1.0"
description = "External Feed Change Detector"
requires-python = ">=3.9"
dependencies = ["PyYAML>=6.0"]

[project.scripts]
efcd = "efcd.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}
packages = ["efcd", "efcd.feeds"]
